Clean Survey Monkey Data and Write to CSV
This is my first project in Python, written approximately three weeks after first learning any Python. 
It takes in excel files that are given by SurveyMonkey.com consisting of emails that are collected at the aquarium using Survey Monkey app.  This excel file is cleaned up by keeping the necessary First Name and Email Address columns and naming them appropriately, and if there are any emails that do not include at least a '.' or '@' are thrown out, as they are unusable.  It then randomizes the list of emails and first names and splits them into two even groups.  At this point, these two groups are put into separate .CSV files so the research assistants can further clean them for mistakes, such as '@gmial.com' instead of '@gmail.com'.  This script reduced error dramatically and allowed for the brunt of the time it took to complete this data entry protocol by about 10 minutes.